+++ minecraft/neko/utils/Utils.java
@@ -3098,7 +3098,7 @@ public class Utils {
             	if (i==44) 
             		AutoClic.cps=Integer.parseInt(ligne);
             	if (i==45) 
-            		Nuker.nukerRadius=Double.parseDouble(ligne);
+            		Nuker.nukerRadius=Integer.parseInt(ligne);
             	if (i==46) {
             		if (!isLock("--ka random")) 
             			KillAura.random=Boolean.parseBoolean(ligne);
@@ -3549,7 +3549,7 @@ public class Utils {
 	                	if (i==44) 
 	                		AutoClic.cps=Integer.parseInt(ligne);
 	                	if (i==45) 
-	                		Nuker.nukerRadius=Double.parseDouble(ligne);
+	                		Nuker.nukerRadius=Integer.parseInt(ligne);
 	                	if (i==46) {
 	                		if (!isLock("--ka random")) 
 	                			KillAura.random=Boolean.parseBoolean(ligne);
